requirements.txt
python-telegram-bot==13.15
requests>=2.25.1
pathlib
pexpect>=4.8.0